% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{convertGeneListFormat}
\alias{convertGeneListFormat}
\title{convertGeneListFormat}
\usage{
convertGeneListFormat(genes, species = "Hs", from = "SYMBOL", to = "ENTREZID")
}
\arguments{
\item{genes}{a list of gene symbols}

\item{species}{species of the gene symbols. One of "Hs" (human) or "Mm" (mouse)}

\item{from}{gene list format to convert from}

\item{to}{gene list format to convert to}
}
\value{
Returns a data-frame containing the gene symbols converted to new format.
}
\description{
convert gene list format between ENSEMBL ENTREZ and HGNC sybmol
}
