% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{convertHumanGeneList}
\usage{
convertHumanGeneList(genes)
}
\arguments{
\item{genes}{a list of genes to convert}
}
\value{
Returns a list of converted gene symbols.
}
\description{
convert human gene list to mouse gene symbols
}
