% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{findDoubletsByBarcode}
\alias{findDoubletsByBarcode}
\title{findDoubletsByBarcode}
\usage{
findDoubletsByBarcode(obj, threshold = 2)
}
\arguments{
\item{obj}{a Seurat object}

\item{threshold}{number of times barcode pair must be observed for cell to be considered not a doublet}
}
\value{
Returns a Seurat object containing doubletBarcode metadata field
}
\description{
Find cells that have a unique combination of barcodes. Probably doublets
}
