% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{findDoubletsBySimulation}
\alias{findDoubletsBySimulation}
\title{findDoubletsBySimulation}
\usage{
findDoubletsBySimulation(obj, ntop = 1000, mads = 3, seed = 10101)
}
\arguments{
\item{obj}{a Seurat or SingleCellExperiment object}

\item{ntop}{number of top variable genes to use for simulations}

\item{mads}{median absolute deviations to use as threshold for doublet calls}

\item{seed}{random seed to use}
}
\value{
Returns a Seurat object containing doubletBarcode metadata field
}
\description{
Find cells in a Seurat / SCE object that may be doublets based on a few different simulation methods from scds and scran packages
}
