% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-plot-funcs.R
\name{getlegend}
\alias{getlegend}
\title{getlegend}
\usage{
getlegend(plot)
}
\arguments{
\item{plot}{a ggplot2 plot object}
}
\value{
Returns a data-frame containing the the 10X cell ID and the lintraceR DNA barcode ID.
}
\description{
Using the ggpubr package, get the legend of a ggplot2 based plot.
Useful for Seurat::AugmentPlot plots because the legends are lost.
}
