% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-plot-funcs.R
\name{plotAugmentDim}
\alias{plotAugmentDim}
\title{plotAugmentDim}
\usage{
plotAugmentDim(
  obj,
  reduction = "umap",
  label = F,
  label.size = 6,
  group.by = NULL,
  title = NULL,
  dpi = 600,
  pt.size = 1,
  cells.highlight = NULL,
  cols.highlight = "red",
  sizes.highlight = 2,
  split.by = NULL
)
}
\arguments{
\item{obj}{a Seurat object}

\item{reduction}{a dimensional reduction within obj}

\item{label}{Logical. Label active identity classes}

\item{label.size}{pt size of label text}

\item{group.by}{identity class to group dataset by prior to plotting}

\item{title}{plot title}

\item{dpi}{resolution (dots per inch) of raster image}

\item{pt.size}{size of points on plot}

\item{cells.highlight}{vector or named list of cell labels to highlight on plot}

\item{cols.highlight}{vector of colors to highlight cells}

\item{sizes.highlight}{pt size of highlighted cells}

\item{split.by}{an ident class to split plots by}
}
\value{
Returns a rasterised ggplot2 object containing the plot (axes are vectors) and on a second page the plot legend in vector format.
}
\description{
Wrapper around Seurat::AugmentPlot and Seurat::Dimplot
Use in conjunction with pdf() statements
}
