% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-plot-funcs.R
\name{plotAugmentFeature}
\alias{plotAugmentFeature}
\title{plotAugmentFeature}
\usage{
plotAugmentFeature(
  obj,
  reduction = "umap",
  features = NULL,
  label = F,
  label.size = 6,
  group.by = NULL,
  title = NULL,
  dpi = 300,
  pt.size = 1,
  order = T,
  ncol = NULL
)
}
\arguments{
\item{obj}{a Seurat object}

\item{reduction}{a dimensional reduction within obj}

\item{features}{a vector of features to plot}

\item{label}{Logical. Label active identity classes}

\item{label.size}{pt size of label text}

\item{group.by}{identity class to group dataset by prior to plotting}

\item{title}{plot title}

\item{dpi}{resolution (dots per inch) of raster image}

\item{pt.size}{size of points on plot}

\item{order}{order by expression value, largest values are plotted last (on top)}

\item{ncol}{number of columns}
}
\value{
Returns a rasterised ggplot2 object containing the plot (axes are vectors) and on a second page the plot legend in vector format.
}
\description{
Wrapper around Seurat::AugmentPlot and Seurat::Dimplot
Use in conjunction with pdf() statements
}
