% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{plotBarcodesPerCell}
\alias{plotBarcodesPerCell}
\title{plotBarcodesPerCell}
\usage{
plotBarcodesPerCell(obj, samplename = "Seurat.obj")
}
\arguments{
\item{obj}{a Seurat object}

\item{samplename}{sample name}
}
\value{
Returns a plot of the number of barcodes detected per cell
}
\description{
plot the number of barcodes detected per cell.
For use with barcode labelled cells. also useful for single cell CRISPR screens
}
