% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-plot-funcs.R
\name{plotElbow}
\alias{plotElbow}
\title{plotElbow}
\usage{
plotElbow(obj, outdir = NULL, samplename = "sample")
}
\arguments{
\item{obj}{a Seurat object}

\item{outdir}{path to output directory}

\item{samplename}{name of Seurat object}
}
\value{
Returns a rasterised ggplot2 object containing the plot (axes are vectors) and on a second page the plot legend in vector format.
}
\description{
Wrapper around Seurat::ElbowPlot and PCAtools::findElbowPoint
}
