% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{run_edgeR_LRT}
\alias{run_edgeR_LRT}
\title{run_edgeR_LRT}
\usage{
run_edgeR_LRT(counts, group, plots = F)
}
\arguments{
\item{counts}{raw counts from single cell object, rows as features and columns as cells}

\item{group}{which factor in object metadata to use for testing. Should only have 2 levels}

\item{plots}{Logical. print plots}
}
\value{
Returns a Seurat object containing doubletBarcode metadata field
}
\description{
Run edge R for 2 groups Ctrl vs Test using the Likelihood Ratio Test (LRT)
Adapted from Sonesson et al. 2019?? Nat comms
https://github.com/csoneson/conquer_comparison/blob/master/scripts/apply_edgeRLRT.R
This is a simple A vs B test. See other functions for more complex GLMs
}
