% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-funcs.R
\name{runClusterProfiler}
\alias{runClusterProfiler}
\title{runClusterProfiler}
\usage{
runClusterProfiler(
  dge,
  lfc.threshold = 0.2,
  padj.threshold = 0.1,
  OrgDb = org.Mm.eg.db,
  category = "BP",
  sample = "gene-set",
  outdir = NULL
)
}
\arguments{
\item{dge}{a Seurat object}

\item{lfc.threshold}{fold change threshold}

\item{padj.threshold}{adjusted p value threshold}

\item{OrgDb}{org db object for relevant species. org.Mm.eg.db or org.Hs.eg.db}

\item{category}{GO category to test. e.g "BP", "MF", or "CC"}

\item{sample}{name of sample gene set}

\item{outdir}{path to desired output directory}
}
\value{
Returns plots and a table of clusterProfiler results
}
\description{
run clusterProfiler on a set of dge results from Seurat::FindMarkers
}
