% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sc_preprocess}
\alias{sc_preprocess}
\title{Preprocess a DataVolley object before creating a Sportscode file}
\usage{
sc_preprocess(x, cover = TRUE)
}
\arguments{
\item{x}{datavolley: a datvolley object as returned by \link{read_dv}}

\item{cover}{logical: convert the \code{skill} label of any dig that follows a block to "Cover"}
}
\value{
A modified version of \code{x}
}
\description{
Preprocess a DataVolley object before creating a Sportscode file
}
\examples{

x <- read_dv(dv_example_file())
x <- sc_preprocess(x)

}
\seealso{
\code{\link{dv2sc}}
}
